kodlar/requirements.txt
streamlit==1.32.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
tensorflow-cpu==2.16.1
pillow==10.2.0